Create Green light or red light
In this coding challenge you’re asked to write a small web game that recreates the iconic green light / red light game from Squid Game. The game will display a box that changes color between green and red. The goal of the game is to get to 15 clicks on the green box within 15 seconds.

You should write all of your game logic in the GreenLightRedLight component. By default, it should display a button with the text “Start Game”. This button will be used to start the game.

Once the button is pressed, your app should display a box that changes its color on a random basis. Ideally, it should change no sooner that 1s and no later than 2s.

If the users click on the red box, they failed the game and you should display Game Over! message and end the game immediately.

If the users click on the green box, you should increment their score count by 1. If they manage to click the green box 15 times within 15 seconds, they win and you should display You win! message.
https://www.reacterry.com/portal/challenges/green-light--red-light